remove circular dependencies. Move session creation to a separate file.
